contractor: take into account both all contracts and contractset contracts when deciding whether to form new contracts or not
